requirements.txt
amqp==5.1.1
APScheduler==3.6.3
arrow==1.2.3
asgiref @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_edg6nzjrx0/croots/recipe/asgiref_1661287549316/work
async-timeout==4.0.2
billiard==3.6.4.0
blessed==1.19.1
cachetools==4.2.2
celery==5.2.7
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_066uop7vqa/croots/recipe/django_1661292540858/work
django-picklefield==3.1
django-redis==5.2.0
kombu==5.2.4
mysqlclient @ file:///Users/builder/ci_310/mysqlclient_1642534389141/work
packaging==21.3
prompt-toolkit==3.0.32
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
six==1.16.0
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_204sy4e5t2/croot/sqlparse_1668793448973/work
tornado==6.1
tzdata==2022.6
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
